+++ swing/framework-model/src/test/java/is/codion/swing/framework/model/SwingEntityEditModelTest.java
@@ -61,7 +61,6 @@ public class SwingEntityEditModelTest {
     model.includeNull().set(true);
     model.nullItem().set("null");
     assertNotNull(model);
-    assertTrue(employeeEditModel.containsComboBoxModel(Employee.JOB));
     assertEquals(model, employeeEditModel.comboBoxModel(Employee.JOB));
     employeeEditModel.refreshComboBoxModels();
     employeeEditModel.clearComboBoxModels();
@@ -73,7 +72,6 @@ public class SwingEntityEditModelTest {
 
   @Test
   void foreignKeyComboBoxModel() {
-    assertFalse(employeeEditModel.containsComboBoxModel(Employee.DEPARTMENT_FK));
     EntityComboBoxModel model = employeeEditModel.foreignKeyComboBoxModel(Employee.DEPARTMENT_FK);
     assertNotNull(model);
     assertTrue(model.cleared());
@@ -84,6 +82,7 @@ public class SwingEntityEditModelTest {
     employeeEditModel.clearComboBoxModels();
     assertTrue(model.cleared());
     assertTrue(model.items().isEmpty());
+    assertSame(model, employeeEditModel.foreignKeyComboBoxModel(Employee.DEPARTMENT_FK));
   }
 
   @Test